# Documents/ApiSpec/openapi.yaml
{
  "openapi": "3.0.1",
  "info": {
    "title": "API",
    "description": "API Spec f�r den digitalen Reisef�hrer",
    "version": "v1"
  },
  "paths": {
...
    "/api/poireview": {
      "post": {
        "tags": [
          "PoIReview"
        ],
        "summary": "Add a new review for a poi",
        "operationId": "AddPoIReview",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/PostPoIReview"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostPoIReview"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PostPoIReview"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PostPoIReview"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/poireview/{poiId}": {
      "get": {
        "tags": [
          "PoIReview"
        ],
        "summary": "Gets all reviews for the given poi",
        "operationId": "GetPoIReviews",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "poiId",
            "in": "path",
            "description": "The id from the poi",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PoIReview"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "PoIReview"
        ],
        "summary": "Deletes the given review.",
        "operationId": "DeletePoIReview",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "poiId",
            "in": "path",
            "description": "The id from the review to delete",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "PointOfInterest": {
        "required": [
          "averageRating",
          "description",
          "imageUrl",
          "isPromoted",
          "latitude",
          "longitude",
          "name",
          "numberOfRatings",
          "poIID",
          "userName"
        ],
        "type": "object",
        "properties": {
          "poIID": {
            "type": "string",
            "description": "Gets or Sets Id",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Gets or Sets Name"
          },
          "userName": {
            "type": "string",
            "description": "Gets or Sets username"
          },
          "description": {
            "type": "string",
            "description": "Gets or Sets Description"
          },
          "latitude": {
            "type": "number",
            "description": "Gets or Sets Latitude",
            "format": "double"
          },
          "longitude": {
            "type": "number",
            "description": "Gets or Sets Longitude",
            "format": "double"
          },
          "averageRating": {
            "type": "number",
            "description": "Gets or Sets AverageRating",
            "format": "double"
          },
          "numberOfRatings": {
            "type": "integer",
            "description": "Gets or Sets NumberOfRatings",
            "format": "int32"
          },
          "imageUrl": {
            "type": "string",
            "description": "Gets or Sets ImageUrl"
          },
          "isPromoted": {
            "type": "boolean",
            "description": "Defines, whether this poi is promoted or not"
          }
        },
        "additionalProperties": false,
        "description": "Model class for a PointOfInterest"
      },
      "PoIReview": {
        "required": [
          "content",
          "pointOfInterest",
          "poIReviewId",
          "rating",
          "userName"
        ],
        "type": "object",
        "properties": {
          "poIReviewId": {
            "type": "string",
            "description": "Gets or Sets poIReviewId",
            "format": "uuid"
          },
          "pointOfInterest": {
            "$ref": "#/components/schemas/PointOfInterest"
          },
          "content": {
            "type": "string",
            "description": "Gets or Sets Content"
          },
          "rating": {
            "type": "integer",
            "description": "Gets or Sets Rating",
            "format": "int32"
          },
          "userName": {
            "type": "string",
            "description": "Gets or Sets username"
          }
        },
        "additionalProperties": false
      },
...
      "PostPoIReview": {
        "required": [
          "content",
          "poIID",
          "rating",
          "userName"
        ],
        "type": "object",
        "properties": {
          "poIID": {
            "type": "string",
            "description": "Gets or Sets Id",
            "format": "uuid"
          },
          "content": {
            "type": "string",
            "description": "Gets or Sets Content"
          },
          "rating": {
            "type": "integer",
            "description": "Gets or Sets Rating",
            "format": "int32"
          },
          "userName": {
            "type": "string",
            "description": "Gets or Sets username"
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    }
  }
}